+++ api/src/certification/results/application/certification-attestation-route.js
@@ -3,7 +3,7 @@ import Joi from 'joi';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { LOCALE } from '../../../shared/domain/constants.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
-import { certificationAttestationController } from './certification-attestation-controller.js';
+import { certificationController } from './certification-controller.js';
 
 const { FRENCH_SPOKEN, ENGLISH_SPOKEN } = LOCALE;
 
@@ -29,7 +29,7 @@ const register = async function (server) {
             assign: 'hasAuthorizationToAccessAdminScope',
           },
         ],
-        handler: certificationAttestationController.getCertificationPDFAttestationsForSession,
+        handler: certificationController.getCertificationPDFAttestationsForSession,
         plugins: {
           'hapi-swagger': {
             produces: ['application/pdf'],
@@ -37,7 +37,7 @@ const register = async function (server) {
         },
         notes: [
           '- **Route accessible par un user Admin**\n' +
-            "- Récupération des attestations de certification d'une session au format PDF" +
+            "- Récupération des certifications d'une session au format PDF" +
             ' via un id de session et un user id',
         ],
         tags: ['api', 'certifications', 'PDF'],
@@ -60,10 +60,10 @@ const register = async function (server) {
             lang: Joi.string().valid(FRENCH_SPOKEN, ENGLISH_SPOKEN),
           }),
         },
-        handler: certificationAttestationController.getPDFAttestation,
+        handler: certificationController.getPDFAttestation,
         notes: [
           '- **Route accessible par un user authentifié**\n' +
-            '- Récupération des informations d’une attestation de certification au format PDF' +
+            '- Récupération des informations d’une certification au format PDF' +
             ' via un id de certification et un user id',
         ],
         tags: ['api', 'certifications', 'PDF'],
@@ -89,7 +89,7 @@ const register = async function (server) {
             lang: Joi.string().valid(FRENCH_SPOKEN, ENGLISH_SPOKEN),
           }),
         },
-        handler: certificationAttestationController.downloadCertificationAttestationsForDivision,
+        handler: certificationController.downloadCertificationAttestationsForDivision,
         tags: ['api', 'organizations'],
         notes: [
           'Cette route est restreinte aux utilisateurs authentifiés',
