+++ api/src/certification/results/application/certification-controller.js
@@ -1,7 +1,13 @@
+import dayjs from 'dayjs';
+
 import * as requestResponseUtils from '../../../../src/shared/infrastructure/utils/request-response-utils.js';
+import { normalizeAndRemoveAccents } from '../../../shared/infrastructure/utils/string-utils.js';
+import { V3CertificationAttestation } from '../domain/models/V3CertificationAttestation.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as privateCertificateSerializer from '../infrastructure/serializers/private-certificate-serializer.js';
 import * as shareableCertificateSerializer from '../infrastructure/serializers/shareable-certificate-serializer.js';
+import * as certificationAttestationPdf from '../infrastructure/utils/pdf/certification-attestation-pdf.js';
+import * as v3CertificationAttestationPdf from '../infrastructure/utils/pdf/v3-certification-attestation-pdf.js';
 
 const getCertificationByVerificationCode = async function (request, h, dependencies = { requestResponseUtils }) {
   const verificationCode = request.payload.verificationCode;
@@ -33,7 +39,145 @@ const findUserCertifications = async function (request) {
   return privateCertificateSerializer.serialize(privateCertificates, { translate });
 };
 
+const getPDFAttestation = async function (
+  request,
+  h,
+  dependencies = { certificationAttestationPdf, v3CertificationAttestationPdf },
+) {
+  const userId = request.auth.credentials.userId;
+  const certificationCourseId = request.params.certificationCourseId;
+  const { i18n } = request;
+  const { isFrenchDomainExtension } = request.query;
+
+  const attestation = await usecases.getCertificationAttestation({
+    userId,
+    certificationCourseId,
+  });
+
+  if (attestation instanceof V3CertificationAttestation) {
+    const fileName = i18n.__('certification-confirmation.file-name', {
+      deliveredAt: dayjs(attestation.deliveredAt).format('YYYYMMDD'),
+    });
+
+    return h
+      .response(
+        dependencies.v3CertificationAttestationPdf.generate({
+          certificates: [attestation],
+          i18n,
+        }),
+      )
+      .code(200)
+      .header('Content-Disposition', `attachment; filename=${fileName}`)
+      .header('Content-Type', 'application/pdf');
+  }
+
+  const { buffer, fileName } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
+    certificates: [attestation],
+    isFrenchDomainExtension,
+    i18n,
+  });
+
+  return h
+    .response(buffer)
+    .header('Content-Disposition', `attachment; filename=${fileName}`)
+    .header('Content-Type', 'application/pdf');
+};
+
+const getCertificationPDFAttestationsForSession = async function (
+  request,
+  h,
+  dependencies = { certificationAttestationPdf, v3CertificationAttestationPdf },
+) {
+  const { i18n } = request;
+
+  const sessionId = request.params.sessionId;
+  const isFrenchDomainExtension = request.query.isFrenchDomainExtension;
+  const attestations = await usecases.getCertificationAttestationsForSession({
+    sessionId,
+  });
+
+  if (attestations.every((attestation) => attestation instanceof V3CertificationAttestation)) {
+    const translatedFileName = i18n.__('certification-confirmation.file-name', {
+      deliveredAt: dayjs(attestations[0].deliveredAt).format('YYYYMMDD'),
+    });
+
+    return h
+      .response(
+        dependencies.v3CertificationAttestationPdf.generate({
+          certificates: attestations,
+          i18n,
+        }),
+      )
+      .code(200)
+      .header('Content-Disposition', `attachment; filename=session-${sessionId}-${translatedFileName}`)
+      .header('Content-Type', 'application/pdf');
+  }
+
+  const { buffer } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
+    certificates: attestations,
+    isFrenchDomainExtension,
+    i18n,
+  });
+
+  const fileName = `attestation-pix-session-${sessionId}.pdf`;
+  return h
+    .response(buffer)
+    .header('Content-Disposition', `attachment; filename=${fileName}`)
+    .header('Content-Type', 'application/pdf');
+};
+
+const downloadCertificationAttestationsForDivision = async function (
+  request,
+  h,
+  dependencies = { certificationAttestationPdf, v3CertificationAttestationPdf },
+) {
+  const organizationId = request.params.organizationId;
+  const { i18n } = request;
+  const { division, isFrenchDomainExtension } = request.query;
+
+  const attestations = await usecases.findCertificationAttestationsForDivision({
+    organizationId,
+    division,
+  });
+
+  if (attestations.every((attestation) => attestation instanceof V3CertificationAttestation)) {
+    const normalizedDivision = normalizeAndRemoveAccents(division);
+
+    const translatedFileName = i18n.__('certification-confirmation.file-name', {
+      deliveredAt: dayjs(attestations[0].deliveredAt).format('YYYYMMDD'),
+    });
+
+    return h
+      .response(
+        dependencies.v3CertificationAttestationPdf.generate({
+          certificates: attestations,
+          i18n,
+        }),
+      )
+      .code(200)
+      .header('Content-Disposition', `attachment; filename=${normalizedDivision}-${translatedFileName}`)
+      .header('Content-Type', 'application/pdf');
+  }
+
+  const { buffer } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
+    certificates: attestations,
+    isFrenchDomainExtension,
+    i18n,
+  });
+
+  const now = dayjs();
+  const fileName = `${now.format('YYYYMMDD')}_attestations_${division}.pdf`;
+
+  return h
+    .response(buffer)
+    .header('Content-Disposition', `attachment; filename=${fileName}`)
+    .header('Content-Type', 'application/pdf');
+};
+
 const certificationController = {
+  getPDFAttestation,
+  getCertificationPDFAttestationsForSession,
+  downloadCertificationAttestationsForDivision,
   getCertification,
   findUserCertifications,
   getCertificationByVerificationCode,
